finished packet callbacks and rs group processing

+++ transmitter/Src/telemetry.h
@@ -41,8 +41,8 @@ typedef struct __packed
     uint8_t type;
     uint32_t seq;        /* Packet sequence ID */
     uint16_t block_id;   /* identifier within a block -> see processLargePayloadData()*/
-    uint8_t frag_index;  /* 0, 1, 2... */
-    uint16_t frag_total; /* total number of fragments */
+    uint8_t frag_index;  /* 0, 1, 2, 3, 4, 5 */
+    uint8_t frag_total; /* SHARDS PER BLOCK */
     uint8_t reliable;
     uint8_t len;
     uint8_t payload[MAX_PAYLOAD_SIZE];
@@ -53,8 +53,8 @@ typedef struct __packed
 typedef struct
 {
     uint16_t block_id;
-    uint16_t frag_idx;
-    uint16_t frag_total;
+    uint8_t frag_idx;
+    uint8_t frag_total;
     uint16_t len;
 } FragmentMeta_t;
